backend/requirements.txt
# ── Web framework ──────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
python-multipart==0.0.9

# ── Auth ───────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── LLM / Ollama ───────────────────────────────────────────
ollama==0.2.1

# ── Vector memory ──────────────────────────────────────────
chromadb==0.5.0

# ── Structured DB ──────────────────────────────────────────
sqlalchemy==2.0.30
aiosqlite==0.20.0

# ── File parsing ───────────────────────────────────────────
pymupdf==1.24.3
pytesseract==0.3.10
Pillow==10.3.0

# ── Scheduling ─────────────────────────────────────────────
apscheduler==3.10.4

# ── Utilities ──────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
aiofiles==23.2.1
